// src/types.ts
export enum EWindowMode {
  Whiteboard = 1,
  ScreenShare = 2,
}

export enum EWindowSizeMode {
  Maximize = 'MAX',
  Minimize = 'MIN',
}

export enum EEventSource {
  ClassRoomWindow = 'ClassRoomWindow',
  TopToolbarWindow = 'TopToolbarWindow',
}

export enum EPopupType {
  InnerWindow = 'InnerWindow',
  OuterWindow = 'OuterWindow',
}